requirements.txt
pywebview==5.4
pypdf==5.9.0
Pillow==11.3.0